+++ src/test/java/cart/dao/CartProductDaoTest.java
@@ -22,7 +22,7 @@ class JdbcCartProductDaoTest {
     @Autowired
     private NamedParameterJdbcTemplate jdbcTemplate;
 
-    private CartDao cartDao;
+    private JdbcCartDao cartDao;
 
     @BeforeEach
     void init() {
